requirements.txt
requests==2.31.0
python-dotenv==1.0.0
openai==1.12.0
pyaudio==0.2.11
wave
langchain-google-genai==1.0.1
langchain-together==0.1.3
google-auth==2.23.4
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0 